pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quant"
version = "0.1.0"
description = "Quantitative Finance Research & Portfolio Management Platform"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "yfinance>=0.2",
    "riskfolio-lib>=6.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov", "ruff", "mypy", "pre-commit"]
notebooks = ["jupyterlab", "ipywidgets"]

[tool.hatch.build.targets.wheel]
packages = ["src/quant"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "NPY", "PD"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests that download data or run long simulations",
    "integration: marks tests requiring external services",
]
